Eliminate slidingIndex of -1
Currently:

1. slidingIndex is initialized as null
1. When a handle is dragged, slidingIndex is set to the index of the
handle being dragged
1. endSlide sets slidingIndex to -1 before removing event listeners and
firing the change event.
1. An onClick handler on the parent div resets slidingIndex to null

The problem is that if the user drags the handle off the parent div
(e.g. resetting to the min value by dragging far to the left), the
onClick handler will not fire, and this leaves the component with
slidingIndex set to -1 indefinitely.

In this state, updateNewValues() will return, so updates to values from
the parent component (e.g. cancelling filters) will not update the
slider.

This PR eliminates this secondary slidingIndex state, setting to null
no matter what.